+++ CHANGELOG.md
@@ -9,10 +9,12 @@
 - Use a URL routes mapper for the server's endpoints
 - Output the server's access URL when starting
 - Make it possible to use glob-style patterns in includes
+- Show the cache volume and limits in the cache status page
 
 ### Fixed
 
 - Fix the crawler's URL rewriting code when the listening port != 1965
+- Purge feedparser's 'bozo_exception' before caching a feed
 
 ## [1.3.2] - 2024-03-04